matching package classes, convert float to double, add destinations for zones, handle all sp for int and float if null

+++ ProjectX.Entities/Models/Benefit/BenSearchReq.cs
@@ -10,7 +10,7 @@ namespace ProjectX.Entities.Models.Benefit
     {
         public int id { get; set; }
         public string title { get; set; }
-        public float limit { get; set; }
+        public double limit { get; set; }
