+++ ProjectX.Entities/Models/Product/ProdSearchReq.cs
@@ -13,9 +13,9 @@ namespace ProjectX.Entities.Models.Product
         public string description { get; set; }
         public bool is_family { get; set; }
         public bool is_active { get; set; }
-        public float is_deductible { get; set; }
-        public float sports_activities { get; set; }
-        public float additional_benefits { get; set; }
+        public double is_deductible { get; set; }
+        public double sports_activities { get; set; }
+        public double additional_benefits { get; set; }
 
 
     }
